tests: move base.py to helpers/base.py
